+++ lib/app.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_training/view/weather_view/weather_view.dart';
+import 'package:flutter_training/view/weather_view/weather_page.dart';
 
 class App extends StatelessWidget {
   const App({super.key});
@@ -12,7 +12,7 @@ class App extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const WeatherView(),
+      home: const WeatherPage(),
     );
   }
 }
