--- lib/view/component/app_button.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppButton extends StatelessWidget {
-  const AppButton({
-    super.key,
-    required this.title,
-    required this.onPressed,
-  });
-
-  final String title;
-  final VoidCallback onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      onPressed: onPressed,
-      child: Text(title),
-    );
-  }
-}
