+++ lib/view/weather_view/component/weather_forecast.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+
+class WeatherForecast extends StatelessWidget {
+  const WeatherForecast({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+
+    return Column(
+      children: [
+        const AspectRatio(
+          aspectRatio: 1,
+          child: Placeholder(),
+        ),
+        Padding(
+          padding: const EdgeInsets.all(16),
+          child: Row(
+            children: [
+              Expanded(
+                child: Text(
+                  '** ℃',
+                  textAlign: TextAlign.center,
+                  style: textTheme.labelLarge?.copyWith(
+                    color: Colors.blue,
+                  ),
+                ),
+              ),
+              Expanded(
+                child: Text(
+                  '** ℃',
+                  textAlign: TextAlign.center,
+                  style: textTheme.labelLarge?.copyWith(
+                    color: Colors.red,
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
