--- lib/view/weather_view/weather_panel.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WeatherPanel extends StatelessWidget {
-  const WeatherPanel({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const AspectRatio(
-      aspectRatio: 1,
-      child: Placeholder(),
-    );
-  }
-}
